Move connectionStore to a file of its own
